4. Make a form for adding a product to the repository and the logic of this form;
