+++ lesson3/pom.xml
@@ -14,7 +14,7 @@
     <packaging>war</packaging>
 
 
-    <name>spring_course Maven Webapp</name>
+    <name>lesson3</name>
     <!-- FIXME change it to the project's website -->
     <url>http://www.example.com</url>
 
@@ -27,7 +27,6 @@
     </properties>
 
     <dependencies>
-
 
 
         <!-- https://mvnrepository.com/artifact/javax.activation/activation -->
@@ -68,11 +67,11 @@
             <version>4.0.0</version>
             <scope>provided</scope>
         </dependency>
-                <dependency>
-                    <groupId>com.fasterxml.jackson.core</groupId>
-                    <artifactId>jackson-databind</artifactId>
-                    <version>2.11.1</version>
-                </dependency>
+        <dependency>
+            <groupId>com.fasterxml.jackson.core</groupId>
+            <artifactId>jackson-databind</artifactId>
+            <version>2.11.1</version>
+        </dependency>
 
         <dependency>
             <groupId>javax.annotation</groupId>
