+++ lesson3/src/main/java/ru/lazarev/springcourse/comtrollers/AddProductController.java
@@ -1,13 +1,32 @@
 package ru.lazarev.springcourse.comtrollers;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import ru.lazarev.springcourse.productapplication.Product;
+import ru.lazarev.springcourse.productapplication.ProductRepository;
 
 @Controller
+@RequestMapping("/add")
 public class AddProductController {
-    @GetMapping("/add")
-    public String homePage() {
+    private final ProductRepository repository;
+
+    @Autowired
+    public AddProductController(ProductRepository repository) {
+        this.repository = repository;
+    }
+
+    @RequestMapping( method = RequestMethod.GET)
+    public String getForm(Model uiModel) {
+        uiModel.addAttribute("product", new Product());
+        return "products/add_products";
+    }
+
+    @RequestMapping(value = "form", method = RequestMethod.POST)
+    public String create(Product product) {
+        repository.addProduct(product);
         return "products/add_products";
     }
 }
