+++ lesson3/src/main/java/ru/lazarev/springcourse/comtrollers/AllProductController.java
@@ -4,8 +4,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import ru.lazarev.springcourse.product_application.Product;
-import ru.lazarev.springcourse.product_application.bean.ProductRepository;
+import ru.lazarev.springcourse.productapplication.Product;
+import ru.lazarev.springcourse.productapplication.ProductRepository;
 
 import java.util.List;
 
@@ -18,16 +18,10 @@ public class AllProductController {
         this.repository = repository;
     }
 
-
     @GetMapping("/all")
     public String allPage(Model model) {
-//        System.out.println(repository.toString());
-
-        System.out.println(repository.getList());
-
         List<Product> product = repository.getList();
-//        model.addAttribute ("repository", repository);
-        model.addAttribute ("product", product);
+        model.addAttribute("product", product);
         return "products/products";
     }
 }
