+++ lesson3/src/main/java/ru/lazarev/springcourse/comtrollers/MainController.java
@@ -7,6 +7,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 public class MainController {
     @GetMapping("/")
     public String homePage() {
-        return "my/main";
+        return "home/main";
     }
 }
