--- lesson3/src/main/java/ru/lazarev/springcourse/product_application/Application.java
@@ -1,18 +1,0 @@
-package ru.lazarev.springcourse.product_application;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * Homework for lesson #3
- *
- * @author Valeriy Lazarev
- * @since 05.03.2021
- */
-
-//@SpringBootApplication
-public class Application {
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class,args);
-    }
-}
