+++ lesson3/src/main/java/ru/lazarev/springcourse/productapplication/ProductRepository.java
@@ -1,15 +1,11 @@
-package ru.lazarev.springcourse.product_application.bean;
+package ru.lazarev.springcourse.productapplication;
 
 import org.springframework.stereotype.Component;
-import ru.lazarev.springcourse.product_application.Product;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-
-/*2. ProductRepository позволяет получить весь список или один товар по id. Создаем бин Cart,
-в который можно добавлять и удалять товары по id.*/
 @Component
 public class ProductRepository {
     private List<Product> list = new ArrayList<>();
@@ -18,16 +14,8 @@ public class ProductRepository {
         return list;
     }
 
-    public Product getProductById(int index) {
-        return this.list.get(index);
-    }
-
     public void addProduct(Product product) {
         this.list.add(product);
-    }
-
-    public int getSize() {
-        return this.list.size();
     }
 
     @Override
