+++ lesson3/src/main/webapp/WEB-INF/views/home/main.html
