+++ lesson3/src/main/webapp/WEB-INF/views/products/products.html
@@ -4,6 +4,7 @@
     <title>All products</title>
 </head>
 <body>
+<br><a href="/">Main</a>
 <h1>All products</h1>
 <h4 th:each="product : ${product}" th:text="${product.getName}"></h4>
 </body>
