+++ lesson3/src/main/webapp/WEB-INF/views/xml.html
@@ -1,15 +1,11 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org" lang="en">
-
 <head>
     <meta charset="ISO-8859-1">
-    <title>My app</title>
+    <title>My xml</title>
 </head>
-
 <body>
-
-<p>xml</p>
-
+<h1>xml page</h1>
+<h4 th:text="'text: ' + ${xmlString}" ></h4>
 </body>
-
 </html>
